fix: split test_global_dispatch_coordinator to comply with LOC limits
Split test_global_dispatch_coordinator.py (778 lines) into focused modules:
- test_dispatch_queue_operations.py (437 lines): Queue management tests
- test_dispatch_state_transitions.py (421 lines): State transitions and logging tests
- test_global_dispatch_coordinator.py (28 lines): Compatibility import point

Removed 500-line exception from config/v3/loc_limits.yaml as all files now
comply with the standard limit.

All 30 tests continue to pass, maintaining full test coverage.
